Pause Added, Improve Boss Follow
Any object that moves/rotates now has Time.deltaTime as a variable in their movement function. DOing so we can manipulate the Time Scale var in the Time class from 0 -1 to control movements in the game, No movements = game paused. We might have to consider any effects this could for any future or current UI components that will use Time as well.