Fixing Local Game Bugs
Had to reformat logic for local player selection and local. Also in a attempt to more easily construct on online version of the game, there will now be separate prefabs of both characters for online and local use.  New scripts dealing with characters online functionality will be created on if-needed basis.